cambio modo api...

+++ database/migrations/20210310095438-create-transaction.js
@@ -27,7 +27,7 @@ module.exports = {
         }
       },
       perception: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(3,1)
       },
       createdAt: {
         allowNull: false,
